Intial README and dump of files
Basic functionality to standup a VM. The 'root' user folder is
mounted via sync folders, as d-g wants to store in there an 8GB
swap file which would use most of the 10GB disk space avaialable.

Shell provisioning is used to install basic devstack dependencies
and read the REPRODUCE_SCRIPT env variable. If set, it must point
to the full URL of the script, which is deployed under /root.

A git cache folder is mounted as well into the VM. If git repos
are available in there, d-g will use those instead of cloning from
git.openstack.org. All copies of the repos are then refreshed by
d-g to the ref needed for the gate.