Start using extra disk for /opt
Mounting /root did not work because the swapfile would be owned by
a non-root user which is not ok. The best option, without modifying the
vagrant box, is to add an extra disk and mount it on /opt which is what
would take most space (apart from the swap file) on a dsvm machine.

Customise the Vagrant file to add a 10GB disk. Extend setup.sh to partition
the disk and create a file system and mount it under /opt.

Update .gitignore and remove root.