Use a large VDI disk for swap and opt
Even using a disk for /opt is not enough to prevent the disk from
filling up, since the swap disk size + OS is too close to the
vagrant box disk size of 10GB.

d-g looks for a device labelled 'ephemeral0', and if found it will
try to partition it and store swap and /opt in there, so we
extend the vdi disk to 20GB and create a fs labelled 'ephemeral0'
in there for d-g to use.

This means that the VM internal disk is mostly unused, but it's
the only option left without hacking into d-g, which I may do in
future.