Fix the HOST_IP code. Add a new dhcp private net.
The HOST_IP code was not doing the right thing, so fixing it.
Adding the host only net to the VM. The idea is that should
be used by the devstack cloud as a provider network, but let's
see if that works.

The provisioning script will have to be adapted to discover
the IP on the new network and the range available, and feed
that info into devstack so that devstack can use it to setup
a pool of floating IPs routable from the host.